refactor: 폴더 명 수정 및 에러 수정
폴더 명 수정 및 에러수정
