+++ src/stores/features/counterSlice.ts
